requirements.txt
args==0.1.0
beautifulsoup4==4.4.1
clint==0.5.1
coloredlogs==5.0
Distance==0.1.3
humanfriendly==1.44.3
Pillow==3.1.1
pymongo==3.2.1
requests==2.9.1
wheel==0.29.0
